solucion bug busqueda y unidades decimales

+++ src/main/java/com/inte_soft/gestionconsumibles/dto/ConsumiblesDto.java
@@ -17,12 +17,12 @@ public class ConsumiblesDto {
     private String referencia;
     private String marca;
     private String unidad;
-    private long cantidad;
+    private double cantidad;
 
     public ConsumiblesDto() {
     }
 
-    public ConsumiblesDto(int ot, String codigo, String descripcion, String tipo, String referencia, String marca, String unidad, long cantidad) {
+    public ConsumiblesDto(int ot, String codigo, String descripcion, String tipo, String referencia, String marca, String unidad, double cantidad) {
         this.ot = ot;
         this.codigo = codigo;
         this.descripcion = descripcion;
@@ -89,11 +89,11 @@ public class ConsumiblesDto {
         this.unidad = unidad;
     }
 
-    public long getCantidad() {
+    public double getCantidad() {
         return cantidad;
     }
 
-    public void setCantidad(long cantidad) {
+    public void setCantidad(double cantidad) {
         this.cantidad = cantidad;
     }
